backend/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
backoff==1.10.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cyrtranslit==1.0
fastapi==0.63.0
idna==2.10
joblib==1.0.1
multidict==5.1.0
nltk==3.5
pydantic==1.8.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
selenium==3.141.0
six==1.15.0
starlette==0.13.6
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
vkaudiotoken==0.5.1
yarl==1.6.3
